+++ library/ui-strings/src/main/res/values/strings.xml
@@ -2563,7 +2563,7 @@
     <string name="direct_room_encryption_enabled_tile_description">Messages in this chat are end-to-end encrypted.</string>
     <string name="direct_room_encryption_enabled_tile_description_future">Messages in this chat will be end-to-end encrypted.</string>
     <string name="direct_room_encryption_enabled_waiting_users">Waiting for users to join ${app_name}</string>
-    <string name="direct_room_encryption_enabled_waiting_users_tile_description">Once all invited users have joined ${app_name}, you will be able to chat and the room will be end to end encrypted</string>
+    <string name="direct_room_encryption_enabled_waiting_users_tile_description">Once invited users have joined ${app_name}, you will be able to chat and the room will be end-to-end encrypted</string>
     <string name="encryption_not_enabled">Encryption not enabled</string>
     <string name="encryption_misconfigured">Encryption is misconfigured</string>
     <string name="encryption_unknown_algorithm_tile_description">The encryption used by this room is not supported</string>
