+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/provider/RoomDisplayNameFallbackProvider.kt
@@ -26,7 +26,7 @@ package org.matrix.android.sdk.api.provider
  */
 interface RoomDisplayNameFallbackProvider {
     fun getNameForRoomInvite(): String
-    fun getNameForEmptyRoom(isDirect: Boolean, directUserId: String?, leftMemberNames: List<String>): String
+    fun getNameForEmptyRoom(isDirect: Boolean, leftMemberNames: List<String>): String
     fun getNameFor1member(name: String): String
     fun getNameFor2members(name1: String, name2: String): String
     fun getNameFor3members(name1: String, name2: String, name3: String): String
