+++ vector/src/main/java/im/vector/app/features/userdirectory/UserListController.kt
@@ -25,16 +25,13 @@ import im.vector.app.R
 import im.vector.app.core.epoxy.errorWithRetryItem
 import im.vector.app.core.epoxy.loadingItem
 import im.vector.app.core.epoxy.noResultItem
-import im.vector.app.core.epoxy.profiles.notifications.TextHeaderItem
 import im.vector.app.core.epoxy.profiles.notifications.textHeaderItem
 import im.vector.app.core.error.ErrorFormatter
-import im.vector.app.core.extensions.isEmail
 import im.vector.app.core.resources.ColorProvider
 import im.vector.app.core.resources.StringProvider
 import im.vector.app.core.ui.list.genericPillItem
 import im.vector.app.features.displayname.getBestName
 import im.vector.app.features.home.AvatarRenderer
-import im.vector.app.features.home.room.detail.timeline.item.messageTextItem
 import im.vector.lib.core.utils.epoxy.charsequence.toEpoxyCharSequence
 import me.gujun.android.span.span
 import org.matrix.android.sdk.api.session.Session
