+++ vector/src/main/java/im/vector/app/features/userdirectory/UserListFragmentArgs.kt
@@ -26,7 +26,6 @@ data class UserListFragmentArgs(
         val submitMenuItemId: Int,
         val excludedUserIds: Set<String>? = null,
         val singleSelection: Boolean = false,
-        val isE2EByDefault: Boolean = false,
         val showInviteActions: Boolean = true,
         val showContactBookAction: Boolean = true,
         val showToolbar: Boolean = true
