+++ vector/src/main/java/im/vector/app/features/userdirectory/UserListViewState.kt
@@ -32,7 +32,7 @@ data class UserListViewState(
         val pendingSelections: Set<PendingSelection> = emptySet(),
         val searchTerm: String = "",
         val singleSelection: Boolean,
-        val isE2EByDefault: Boolean,
+        val isE2EByDefault: Boolean = false,
         val configuredIdentityServer: String? = null,
         private val showInviteActions: Boolean,
         val showContactBookAction: Boolean
@@ -41,7 +41,6 @@ data class UserListViewState(
     constructor(args: UserListFragmentArgs) : this(
             excludedUserIds = args.excludedUserIds,
             singleSelection = args.singleSelection,
-            isE2EByDefault = args.isE2EByDefault,
             showInviteActions = args.showInviteActions,
             showContactBookAction = args.showContactBookAction
     )
